Dev feature game policies

+++ app/Http/Controllers/GameController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Resources\GameGroupedByPlayerResource;
 use App\Http\Resources\GameResource;
 use App\Models\Game;
 use App\Models\User;
 use App\Services\StatisticsService;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 
 class GameController extends Controller
 {
@@ -21,9 +21,16 @@ class GameController extends Controller
    */
   public function adminIndex(): JsonResponse
   {
-    $games = Game::all();
-    return response()->json(GameGroupedByPlayerResource::collection($games), 200);
+    Gate::authorize('viewAny', User::class);
+    $gamesGroupedByPlayer = Game::getGamesGroupedByPlayer();
+
+    if (empty($gamesGroupedByPlayer)) {
+      return response()->json(['message' => 'No games found'], 404);
+    }
+
+    return response()->json($gamesGroupedByPlayer, 200);
   }
+
 
   /**
    * @lrd:start
@@ -83,14 +90,16 @@ class GameController extends Controller
   /**
    * @lrd:start
    * # Delete Games
-   * Remove all games of the specified user from storage.
+   * Remove all games of the authenticated user.
    *
    * @param User $user User whose games are to be deleted.
    * @return JsonResponse Empty response with status code 204 on success.
    * @lrd:end
    */
-  public function destroy(User $user): JsonResponse
+  public function destroy(): JsonResponse
   {
+    $user = auth()->user();
+    Gate::authorize('delete', $user);
     Game::where('user_id', $user->id)->delete();
 
     return response()->json(null, 204);
